Move render/shm_format functions to render/pixel_format
