+++ include/render/pixel_format.h
@@ -20,4 +20,7 @@ struct wlr_pixel_format_info {
 
 const struct wlr_pixel_format_info *drm_get_pixel_format_info(uint32_t fmt);
 
+uint32_t convert_wl_shm_format_to_drm(enum wl_shm_format fmt);
+enum wl_shm_format convert_drm_format_to_wl_shm(uint32_t fmt);
+
 #endif
