+++ render/meson.build
@@ -4,7 +4,6 @@ wlr_files += files(
 	'egl.c',
 	'drm_format_set.c',
 	'gbm_allocator.c',
-	'shm_format.c',
 	'pixel_format.c',
 	'swapchain.c',
 	'wlr_renderer.c',
