--- render/shm_format.c
@@ -1,24 +1,0 @@
-#include <drm_fourcc.h>
-#include "render/shm_format.h"
-
-uint32_t convert_wl_shm_format_to_drm(enum wl_shm_format fmt) {
-	switch (fmt) {
-	case WL_SHM_FORMAT_XRGB8888:
-		return DRM_FORMAT_XRGB8888;
-	case WL_SHM_FORMAT_ARGB8888:
-		return DRM_FORMAT_ARGB8888;
-	default:
-		return (uint32_t)fmt;
-	}
-}
-
-enum wl_shm_format convert_drm_format_to_wl_shm(uint32_t fmt) {
-	switch (fmt) {
-	case DRM_FORMAT_XRGB8888:
-		return WL_SHM_FORMAT_XRGB8888;
-	case DRM_FORMAT_ARGB8888:
-		return WL_SHM_FORMAT_ARGB8888;
-	default:
-		return (enum wl_shm_format)fmt;
-	}
-}
